requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
requests==2.31.0
python-dotenv==1.0.0

# Databricks Apps 배포 시 추가 권장 패키지
gunicorn==21.2.0

